chore(concurrent-keys): remove project scaffolding before v8 merge
Removes concurrent-keys project scaffolding from the feature branch ahead of the
v8 merge, leaving only the shipped feature code and its regression tests.

- Delete .agent-docs/concurrent-keys/ (design, plan, T0 PoC findings). Preserved
  off-branch in Confluence for reference.
- Delete the re-anchor PoC test (env_context_reanchor_test.go). Its shared test
  helpers are extracted into env_context_reanchor_helpers_test.go so the surviving
  re-anchor regression tests still compile; the PoC hypotheses are already covered
  by the synchronous/double/rollback/bigsegment tests.
- Delete the harness reference test and the now-orphaned archive fixture builder
  (its only consumer was that reference test).
- Scrub task IDs, design-document section references, and PoC/spike language from
  comments in the surviving files.
