--- internal/relayenv/env_context_reanchor_test.go
@@ -1,659 +1,0 @@
-package relayenv
-
-// Re-anchoring proof-of-concept tests.
-//
-// These tests validate the upstream SDK-client swap mechanism that the re-anchor implementation builds
-// on. Each test answers one of the seven hypotheses about how re-anchoring should behave. They are
-// written as durable, executable probes of today's primitives so they survive as regression tests and
-// as the executable spec for the re-anchor implementation.
-//
-// Terminology: "re-anchor" = swapping the single upstream SDK client when sdkKey.value changes.
-// Today there is no dedicated re-anchor method; the closest existing path is ReconcileCredentials with
-// an expiring (grace-period) key (which rotates the primary SDK key and stands up a new client), so
-// several tests drive that path and observe where it falls short of the requirements.
-
-import (
-	"errors"
-	"net/http"
-	"sync"
-	"testing"
-	"time"
-
-	"github.com/launchdarkly/ld-relay/v8/config"
-	"github.com/launchdarkly/ld-relay/v8/internal/basictypes"
-	"github.com/launchdarkly/ld-relay/v8/internal/bigsegments"
-	"github.com/launchdarkly/ld-relay/v8/internal/credential"
-	"github.com/launchdarkly/ld-relay/v8/internal/httpconfig"
-	"github.com/launchdarkly/ld-relay/v8/internal/sdks"
-	st "github.com/launchdarkly/ld-relay/v8/internal/sharedtest"
-	"github.com/launchdarkly/ld-relay/v8/internal/sharedtest/testclient"
-	"github.com/launchdarkly/ld-relay/v8/internal/store"
-	"github.com/launchdarkly/ld-relay/v8/internal/streams"
-	"github.com/launchdarkly/ld-relay/v8/internal/util"
-
-	"github.com/launchdarkly/eventsource"
-	"github.com/launchdarkly/go-sdk-common/v3/ldlog"
-	"github.com/launchdarkly/go-sdk-common/v3/ldlogtest"
-	ld "github.com/launchdarkly/go-server-sdk/v7"
-	"github.com/launchdarkly/go-server-sdk/v7/ldcomponents"
-	"github.com/launchdarkly/go-server-sdk/v7/subsystems"
-	"github.com/launchdarkly/go-server-sdk/v7/subsystems/ldstoreimpl"
-	"github.com/launchdarkly/go-server-sdk/v7/subsystems/ldstoretypes"
-	helpers "github.com/launchdarkly/go-test-helpers/v3"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-// reanchorTestKey2 is the "new anchor" SDK key we re-anchor onto. It is deliberately NOT in the real
-// sdk-<uuid> credential format so it won't trip secret scanners; relay treats SDK keys as opaque
-// non-empty strings, so any value works here.
-const reanchorTestKey2 = config.SDKKey("reanchor-poc-new-anchor")
-
-// recordingStreamUpdates is a streams.EnvStreamUpdates that counts the broadcasts it receives, so we
-// can observe whether a re-anchor produces duplicate downstream "put"s.
-type recordingStreamUpdates struct {
-	mu             sync.Mutex
-	allDataUpdates int
-	singleUpdates  int
-	invalidations  int
-}
-
-func (r *recordingStreamUpdates) SendAllDataUpdate(_ []ldstoretypes.Collection) {
-	r.mu.Lock()
-	r.allDataUpdates++
-	r.mu.Unlock()
-}
-
-func (r *recordingStreamUpdates) SendSingleItemUpdate(_ ldstoretypes.DataKind, _ string, _ ldstoretypes.ItemDescriptor) {
-	r.mu.Lock()
-	r.singleUpdates++
-	r.mu.Unlock()
-}
-
-func (r *recordingStreamUpdates) InvalidateClientSideState() {
-	r.mu.Lock()
-	r.invalidations++
-	r.mu.Unlock()
-}
-
-func (r *recordingStreamUpdates) allDataCount() int {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	return r.allDataUpdates
-}
-
-// sharedStoreFactory is a DataStore configurer that hands back the SAME underlying store instance on
-// every Build call. It models a persistent store (Redis/DynamoDB), where the data lives outside the
-// process and survives the recreation of the wrapping store on a client swap.
-type sharedStoreFactory struct {
-	store subsystems.DataStore
-}
-
-func (f *sharedStoreFactory) Build(_ subsystems.ClientContext) (subsystems.DataStore, error) {
-	return f.store, nil
-}
-
-// reanchor re-anchors env onto newKey while keeping oldKey accepted for a grace hour (the old
-// client stays up while the new one is built, then closes when the commit lands). This mirrors the
-// backend's default-rotation behavior: the new anchor is non-expiring, the demoted old anchor
-// carries an expiry. It drives the time-injectable reconcileCredentials directly so the
-// grace-period math is deterministic.
-func reanchor(t *testing.T, env EnvContext, newKey, oldKey config.SDKKey, now time.Time) {
-	t.Helper()
-	set, err := credential.NewAcceptedSetBuilder().
-		WithAnchor(credential.SDKKeyParams{Value: newKey}).
-		WithSDKKey(credential.SDKKeyParams{Value: oldKey, Expiry: util.PtrOrNil(now.Add(time.Hour))}).
-		Build()
-	require.NoError(t, err)
-	env.(*envContextImpl).reconcileCredentials(set, now)
-}
-
-// -----------------------------------------------------------------------------------------------
-// Hypothesis 1: Two SDK clients sharing a storeAdapter don't corrupt store invariants.
-// -----------------------------------------------------------------------------------------------
-
-func TestReanchorPoC_H1_SharedStoreAdapterRebuildSemantics(t *testing.T) {
-	featureKind := ldstoreimpl.Features()
-	flagKey := st.Flag1ServerSide.Flag.Key
-
-	// Original PoC finding: each storeAdapter.Build call constructed a fresh wrapper around a fresh
-	// underlying store, so the new anchor's client would start empty. The store-handover change
-	// (SSERelayDataStoreAdapter.Build reusing its existing wrapper, with refcounted Close) inverts
-	// this: the second client receives the SAME wrapper, with its data still in place. This sub-test
-	// now asserts the post-fix invariant.
-	t.Run("in-memory factory reuses the existing store on a second client init (store handover)", func(t *testing.T) {
-		rec := &recordingStreamUpdates{}
-		adapter := store.NewSSERelayDataStoreAdapter(ldcomponents.InMemoryDataStore(), rec)
-
-		// First client init populates the store, as the original anchor's client would.
-		s1, err := adapter.Build(subsystems.BasicClientContext{})
-		require.NoError(t, err)
-		require.NoError(t, s1.Init(st.AllData))
-		require.Same(t, s1, adapter.GetStore())
-
-		got, err := adapter.GetStore().Get(featureKind, flagKey)
-		require.NoError(t, err)
-		require.NotNil(t, got.Item, "data should be present after the first client's sync")
-
-		// Second client init = the re-anchor's "start new client" step.
-		s2, err := adapter.Build(subsystems.BasicClientContext{})
-		require.NoError(t, err)
-
-		// Post-fix: Build hands the existing wrapper to the new client and the adapter still points at it.
-		assert.Same(t, s1, s2, "store handover: Build returns the existing wrapper")
-		assert.Same(t, s2, adapter.GetStore(), "the adapter still points at the shared wrapper")
-
-		// The wrapper stays initialized and the data survives — no empty-store window for the new anchor.
-		assert.True(t, adapter.GetStore().IsInitialized(), "the shared store remains initialized")
-		got2, err := adapter.GetStore().Get(featureKind, flagKey)
-		require.NoError(t, err)
-		assert.NotNil(t, got2.Item, "data persists across handover")
-	})
-
-	// With a persistent store, the underlying data lives outside the wrapper, so the swap preserves it.
-	// This is the configuration in which the "shared store" assumption actually holds.
-	t.Run("shared (persistent) underlying store preserves data across client init", func(t *testing.T) {
-		underlying, err := ldcomponents.InMemoryDataStore().Build(subsystems.BasicClientContext{})
-		require.NoError(t, err)
-		rec := &recordingStreamUpdates{}
-		adapter := store.NewSSERelayDataStoreAdapter(&sharedStoreFactory{store: underlying}, rec)
-
-		s1, err := adapter.Build(subsystems.BasicClientContext{})
-		require.NoError(t, err)
-		require.NoError(t, s1.Init(st.AllData))
-
-		// Re-anchor's "start new client" step.
-		s2, err := adapter.Build(subsystems.BasicClientContext{})
-		require.NoError(t, err)
-
-		// The wrapper is new, but it wraps the SAME underlying store, so data + initialization survive.
-		assert.True(t, s2.IsInitialized(), "persistent store stays initialized across the swap")
-		got, err := s2.Get(featureKind, flagKey)
-		require.NoError(t, err)
-		assert.NotNil(t, got.Item, "data survives the swap when the underlying store is shared")
-	})
-}
-
-// -----------------------------------------------------------------------------------------------
-// Hypothesis 2: Downstream SSE connections tolerate the swap.
-//   2a: an open downstream connection survives a re-anchor and keeps receiving events.
-//   2b: the new anchor's initial sync re-broadcasts a (duplicate) "put" downstream.
-// -----------------------------------------------------------------------------------------------
-
-func TestReanchorPoC_H2_DownstreamConnectionSurvivesReAnchor(t *testing.T) {
-	envConfig := st.EnvClientSide.Config
-
-	fakeBigSegmentStoreFactory := func(config.EnvConfig, config.Config, ldlog.Loggers) (bigsegments.BigSegmentStore, error) {
-		return bigsegments.NewNullBigSegmentStore(), nil
-	}
-	fakeSynchronizerFactory := &mockBigSegmentSynchronizerFactory{}
-
-	mockLog := ldlogtest.NewMockLog()
-	defer mockLog.DumpIfTestFailed(t)
-
-	jsClientStreams := streams.NewStreamProvider(basictypes.JSClientPingStream, time.Hour, 0)
-	clientCh := make(chan *testclient.FakeLDClient, 10)
-	sdkStartedCh := make(chan EnvContext, 10)
-	env, err := NewEnvContext(EnvContextImplParams{
-		Identifiers:                   EnvIdentifiers{ConfiguredName: st.EnvClientSide.Name},
-		EnvConfig:                     envConfig,
-		AllConfig:                     config.Config{},
-		BigSegmentStoreFactory:        fakeBigSegmentStoreFactory,
-		BigSegmentSynchronizerFactory: fakeSynchronizerFactory.create,
-		ClientFactory:                 testclient.FakeLDClientFactoryWithChannel(true, clientCh),
-		SDKBigSegmentsConfigFactory: ldcomponents.BigSegments(
-			st.ExistingInstance[subsystems.BigSegmentStore](&st.NoOpSDKBigSegmentStore{}),
-		),
-		StreamProviders:  []streams.StreamProvider{jsClientStreams},
-		ConnectionMapper: mockConnectionMapper{},
-		Loggers:          mockLog.Loggers,
-	}, sdkStartedCh)
-	require.NoError(t, err)
-	defer env.Close()
-
-	synchronizer := fakeSynchronizerFactory.synchronizer
-	require.NotNil(t, synchronizer)
-
-	// Wait for the original anchor client and initialize the store so the client-side stream is ready.
-	<-sdkStartedCh
-	require.NoError(t, env.GetStore().Init(nil))
-
-	streamHandler := env.GetStreamHandler(jsClientStreams, envConfig.EnvID)
-	req, _ := http.NewRequest("GET", "", nil)
-	st.WithStreamRequest(t, req, streamHandler, func(eventCh <-chan eventsource.Event) {
-		initEvent := helpers.RequireValue(t, eventCh, time.Minute)
-		assert.Equal(t, "ping", initEvent.Event())
-		if !helpers.AssertNoMoreValues(t, eventCh, 100*time.Millisecond) {
-			t.FailNow()
-		}
-
-		// --- Re-anchor while the downstream connection is open. ---
-		// The connection is keyed on the env ID (a ScopedCredential), independent of the upstream SDK
-		// key, so swapping the SDK anchor must not disturb it.
-		start := time.Unix(1000, 0)
-		reanchor(t, env, reanchorTestKey2, envConfig.SDKKey, start)
-
-		// The new anchor client comes up on a background goroutine. Credential additions start the client
-		// with a nil readyCh (so it does NOT signal sdkStartedCh); wait on the credential set instead.
-		require.Eventually(t, func() bool {
-			creds := env.GetCredentials()
-			for _, c := range creds {
-				if c == reanchorTestKey2 {
-					return true
-				}
-			}
-			return false
-		}, time.Second, 10*time.Millisecond, "rotation to the new anchor should be applied")
-
-		// FINDING: the open client-side connection survives the swap and still delivers events.
-		// T2.d re-anchors the big-segment synchronizer, so a post-re-anchor update arrives on the CURRENT
-		// (rebuilt) synchronizer, not the retired one (whose channel is now closed).
-		current := fakeSynchronizerFactory.synchronizer
-		require.NotSame(t, synchronizer, current, "the synchronizer was rebuilt on re-anchor")
-		current.updateCh <- bigsegments.UpdatesSummary{SegmentKeysUpdated: []string{"fake-segment-key"}}
-		pingEvent := helpers.RequireValue(t, eventCh, time.Second)
-		assert.Equal(t, "ping", pingEvent.Event(), "downstream connection should survive the re-anchor")
-	})
-}
-
-func TestReanchorPoC_H2_NewClientInitialSyncRebroadcastsPut(t *testing.T) {
-	rec := &recordingStreamUpdates{}
-	adapter := store.NewSSERelayDataStoreAdapter(ldcomponents.InMemoryDataStore(), rec)
-
-	// Original anchor client builds and performs its initial sync -> one downstream "put".
-	s1, err := adapter.Build(subsystems.BasicClientContext{})
-	require.NoError(t, err)
-	require.NoError(t, s1.Init(st.AllData))
-	require.Equal(t, 1, rec.allDataCount())
-
-	// Re-anchor: the new anchor client builds a fresh store and performs its OWN initial sync.
-	s2, err := adapter.Build(subsystems.BasicClientContext{})
-	require.NoError(t, err)
-	require.NoError(t, s2.Init(st.AllData))
-
-	// FINDING: the new anchor's initial sync produces a second full "put" to every connected downstream
-	// stream. From a downstream SDK's perspective this is a duplicate put. It is tolerable (SDKs apply
-	// puts idempotently) but the re-anchor implementation must expect it; it is not a corruption.
-	assert.Equal(t, 2, rec.allDataCount(), "the new anchor's initial sync re-broadcasts a full put")
-}
-
-// -----------------------------------------------------------------------------------------------
-// Hypothesis 3: Big-segment sync after re-anchor.
-// -----------------------------------------------------------------------------------------------
-
-// capturingBigSegmentSynchronizerFactory records the SDK key it was constructed with and how many
-// times it was invoked, so we can detect whether a re-anchor re-wires big-segment sync.
-type capturingBigSegmentSynchronizerFactory struct {
-	mu           sync.Mutex
-	createCount  int
-	lastSDKKey   config.SDKKey
-	synchronizer *mockBigSegmentSynchronizer
-}
-
-func (f *capturingBigSegmentSynchronizerFactory) create(
-	_ httpconfig.HTTPConfig,
-	_ bigsegments.BigSegmentStore,
-	_ string,
-	_ string,
-	_ config.EnvironmentID,
-	sdkKey config.SDKKey,
-	_ ldlog.Loggers,
-	_ string,
-) bigsegments.BigSegmentSynchronizer {
-	f.mu.Lock()
-	defer f.mu.Unlock()
-	f.createCount++
-	f.lastSDKKey = sdkKey
-	f.synchronizer = &mockBigSegmentSynchronizer{updateCh: make(chan bigsegments.UpdatesSummary)}
-	return f.synchronizer
-}
-
-func (f *capturingBigSegmentSynchronizerFactory) snapshot() (int, config.SDKKey) {
-	f.mu.Lock()
-	defer f.mu.Unlock()
-	return f.createCount, f.lastSDKKey
-}
-
-// latest returns the most recently created synchronizer (the current one after a re-anchor rebuild).
-func (f *capturingBigSegmentSynchronizerFactory) latest() *mockBigSegmentSynchronizer {
-	f.mu.Lock()
-	defer f.mu.Unlock()
-	return f.synchronizer
-}
-
-func TestReanchorPoC_H3_BigSegmentSyncFollowsAnchorOnReAnchor(t *testing.T) {
-	envConfig := st.EnvMain.Config
-
-	fakeBigSegmentStoreFactory := func(config.EnvConfig, config.Config, ldlog.Loggers) (bigsegments.BigSegmentStore, error) {
-		return bigsegments.NewNullBigSegmentStore(), nil
-	}
-	capturing := &capturingBigSegmentSynchronizerFactory{}
-
-	mockLog := ldlogtest.NewMockLog()
-	defer mockLog.DumpIfTestFailed(t)
-
-	clientCh := make(chan *testclient.FakeLDClient, 10)
-	env, err := NewEnvContext(EnvContextImplParams{
-		Identifiers:                   EnvIdentifiers{ConfiguredName: st.EnvMain.Name},
-		EnvConfig:                     envConfig,
-		AllConfig:                     config.Config{},
-		BigSegmentStoreFactory:        fakeBigSegmentStoreFactory,
-		BigSegmentSynchronizerFactory: capturing.create,
-		ClientFactory:                 testclient.FakeLDClientFactoryWithChannel(true, clientCh),
-		SDKBigSegmentsConfigFactory: ldcomponents.BigSegments(
-			st.ExistingInstance[subsystems.BigSegmentStore](&st.NoOpSDKBigSegmentStore{}),
-		),
-		ConnectionMapper: mockConnectionMapper{},
-		Loggers:          mockLog.Loggers,
-	}, nil)
-	require.NoError(t, err)
-	defer env.Close()
-
-	count, sdkKey := capturing.snapshot()
-	require.Equal(t, 1, count, "the synchronizer is constructed once at env creation")
-	require.Equal(t, envConfig.SDKKey, sdkKey, "it is wired to the original anchor's SDK key")
-
-	// Re-anchor onto a new SDK key.
-	start := time.Unix(1000, 0)
-	reanchor(t, env, reanchorTestKey2, envConfig.SDKKey, start)
-	require.Eventually(t, func() bool {
-		for _, c := range env.GetCredentials() {
-			if c == reanchorTestKey2 {
-				return true
-			}
-		}
-		return false
-	}, time.Second, 10*time.Millisecond, "rotation to the new anchor should be applied")
-
-	// T2.d: the re-anchor recreates the big-segment synchronizer on the NEW anchor key, so its
-	// poll/stream requests authenticate with the current anchor instead of the retired one. The
-	// synchronizer bakes its SDK key in at construction and is not restartable, so re-anchoring rebuilds
-	// it. (reconcileCredentials -> commitReanchor -> reanchorBigSegmentSync runs synchronously.)
-	count, sdkKey = capturing.snapshot()
-	assert.Equal(t, 2, count, "the synchronizer is recreated on re-anchor")
-	assert.Equal(t, reanchorTestKey2, sdkKey, "the new synchronizer references the new anchor key")
-}
-
-// -----------------------------------------------------------------------------------------------
-// Hypothesis 4: httpconfig stays functional after re-anchor.
-// -----------------------------------------------------------------------------------------------
-
-func TestReanchorPoC_H4_HTTPConfigIsKeyIndependentExceptAuthHeader(t *testing.T) {
-	loggers := ldlog.NewDisabledLoggers()
-	key1 := config.SDKKey("sdk-key-one")
-	key2 := config.SDKKey("sdk-key-two")
-
-	var proxy config.ProxyConfig
-	var httpC config.HTTPConfig
-
-	c1, err := httpconfig.NewHTTPConfig(proxy, httpC, key1, "user-agent", loggers)
-	require.NoError(t, err)
-	c2, err := httpconfig.NewHTTPConfig(proxy, httpC, key2, "user-agent", loggers)
-	require.NoError(t, err)
-
-	// The ONLY key-dependent artifact is the Authorization default header on the pre-built SDK HTTP
-	// config.
-	assert.Equal(t, string(key1), c1.SDKHTTPConfig.DefaultHeaders.Get("Authorization"))
-	assert.Equal(t, string(key2), c2.SDKHTTPConfig.DefaultHeaders.Get("Authorization"))
-
-	// Everything else (proxy settings, user agent, and the rest of the default headers) is identical
-	// and key-independent.
-	h1 := c1.SDKHTTPConfig.DefaultHeaders.Clone()
-	h2 := c2.SDKHTTPConfig.DefaultHeaders.Clone()
-	h1.Del("Authorization")
-	h2.Del("Authorization")
-	assert.Equal(t, h1, h2, "non-auth default headers are key-independent")
-	assert.Equal(t, c1.ProxyConfig, c2.ProxyConfig, "proxy config is key-independent")
-
-	// FINDING: httpconfig needs NO re-wire on re-anchor. Relay injects the *builder*
-	// (SDKHTTPConfigFactory) into ld.Config.HTTP, and the SDK rebuilds the HTTP config with the new
-	// anchor key when it constructs the new client, so the Authorization header is set correctly for the
-	// new anchor automatically. The pre-built SDKHTTPConfig / Client() (used for event + big-segment
-	// transport) is key-independent except for that Authorization header, which those components set per
-	// request from their own credential rather than reading it from httpconfig.
-}
-
-// -----------------------------------------------------------------------------------------------
-// Hypothesis 5: Order of operations / the in-memory store window.
-// -----------------------------------------------------------------------------------------------
-
-// TestReanchorPoC_H5_StoreSurvivesReAnchor was originally a proof-of-concept test asserting the
-// *broken* pre-handover behavior: that re-anchor caused the env's in-memory store to be replaced with
-// a fresh, empty one. Once SSERelayDataStoreAdapter.Build was changed to hand over the existing wrapper
-// to the new client (refcounted Close), that breakage is gone. This now asserts the post-fix
-// invariant — the data store instance survives the re-anchor and keeps its data — and is preserved
-// as the executable proof that handover holds end-to-end through env_context.
-func TestReanchorPoC_H5_StoreSurvivesReAnchor(t *testing.T) {
-	featureKind := ldstoreimpl.Features()
-	flagKey := st.Flag1ServerSide.Flag.Key
-	envConfig := st.EnvMain.Config
-
-	mockLog := ldlogtest.NewMockLog()
-	defer mockLog.DumpIfTestFailed(t)
-
-	clientCh := make(chan *testclient.FakeLDClient, 10)
-	readyCh := make(chan EnvContext, 1)
-	env := makeBasicEnv(t, envConfig, testclient.FakeLDClientFactoryWithChannel(true, clientCh), mockLog.Loggers, readyCh)
-	defer env.Close()
-
-	require.Equal(t, env, requireEnvReady(t, readyCh))
-	client1 := requireClientReady(t, clientCh)
-	require.Eventually(t, func() bool { return env.GetClient() == client1 }, time.Second, 10*time.Millisecond)
-
-	// Populate the store as the original anchor's client would have via its stream sync.
-	require.NoError(t, env.GetStore().Init(st.AllData))
-	oldStore := env.GetStore()
-	got, err := oldStore.Get(featureKind, flagKey)
-	require.NoError(t, err)
-	require.NotNil(t, got.Item)
-
-	// Re-anchor onto a new key (old key kept accepted for a grace hour; the old client serves while
-	// the new one is built and closes once the commit lands -- the "start-new-before-close-old" ordering).
-	start := time.Unix(1000, 0)
-	reanchor(t, env, reanchorTestKey2, envConfig.SDKKey, start)
-
-	client2 := requireClientReady(t, clientCh)
-	require.Eventually(t, func() bool { return env.GetClient() == client2 }, time.Second, 10*time.Millisecond,
-		"GetClient should return the new anchor's client once it is registered")
-
-	// Post-fix: the adapter handed the existing wrapper to the new client, so the env's store is the
-	// same instance, still initialized, and the data is intact. There is no empty-store window for the
-	// new anchor.
-	newStore := env.GetStore()
-	assert.Same(t, oldStore, newStore, "the data store instance survives re-anchor (store handover)")
-	assert.True(t, newStore.IsInitialized(), "the store stays initialized across re-anchor")
-	got2, err := newStore.Get(featureKind, flagKey)
-	require.NoError(t, err)
-	assert.NotNil(t, got2.Item, "data is preserved across re-anchor")
-}
-
-// TestReanchorPoC_H5_StoreHandoverAvoidsEmptyWindow validates the reviewer suggestion that, because
-// relay owns the data store implementation (it hands the SDK a single storeAdapter), the re-anchor can
-// hand the existing store over to the new client instead of letting it build a fresh one. Modeled here
-// by a DataStoreFactory that returns the same underlying store on every Build; the production change
-// is to make SSERelayDataStoreAdapter reuse its store across the swap. With handover the new anchor's
-// client sees the populated, initialized store immediately -- no empty-store window (contrast
-// TestReanchorPoC_H5_InMemoryStoreIsWipedByReAnchor).
-//
-// CAVEAT for the implementation (not reproducible with the fake client, so documented here):
-// streamUpdatesStoreWrapper.Close() closes the underlying store. If the new client wraps the
-// SAME underlying store, closing the retiring client must NOT close it -- the store's lifecycle has to
-// be owned by the adapter, not by the client being retired.
-func TestReanchorPoC_H5_StoreHandoverAvoidsEmptyWindow(t *testing.T) {
-	featureKind := ldstoreimpl.Features()
-	flagKey := st.Flag1ServerSide.Flag.Key
-	envConfig := st.EnvMain.Config
-
-	mockLog := ldlogtest.NewMockLog()
-	defer mockLog.DumpIfTestFailed(t)
-
-	// A store factory that hands the same underlying store to every client (the "handover" model).
-	underlying, err := ldcomponents.InMemoryDataStore().Build(subsystems.BasicClientContext{})
-	require.NoError(t, err)
-	handoverFactory := &sharedStoreFactory{store: underlying}
-
-	clientCh := make(chan *testclient.FakeLDClient, 10)
-	readyCh := make(chan EnvContext, 1)
-	env, err := NewEnvContext(EnvContextImplParams{
-		Identifiers:      EnvIdentifiers{ConfiguredName: envName},
-		EnvConfig:        envConfig,
-		ClientFactory:    testclient.FakeLDClientFactoryWithChannel(true, clientCh),
-		DataStoreFactory: handoverFactory,
-		ConnectionMapper: mockConnectionMapper{},
-		Loggers:          mockLog.Loggers,
-	}, readyCh)
-	require.NoError(t, err)
-	defer env.Close()
-
-	require.Equal(t, env, requireEnvReady(t, readyCh))
-	client1 := requireClientReady(t, clientCh)
-	require.Eventually(t, func() bool { return env.GetClient() == client1 }, time.Second, 10*time.Millisecond)
-
-	// Populate the store as the original anchor's client would have.
-	require.NoError(t, env.GetStore().Init(st.AllData))
-
-	// Re-anchor onto a new key.
-	start := time.Unix(1000, 0)
-	reanchor(t, env, reanchorTestKey2, envConfig.SDKKey, start)
-
-	client2 := requireClientReady(t, clientCh)
-	require.Eventually(t, func() bool { return env.GetClient() == client2 }, time.Second, 10*time.Millisecond)
-
-	// FINDING: with store handover there is no empty-store window -- the new client's store is still
-	// initialized and still holds the data, because the underlying store was reused rather than rebuilt.
-	newStore := env.GetStore()
-	assert.True(t, newStore.IsInitialized(), "handed-over store stays initialized across the re-anchor")
-	got, err := newStore.Get(featureKind, flagKey)
-	require.NoError(t, err)
-	assert.NotNil(t, got.Item, "data is preserved across the re-anchor when the store is handed over")
-}
-
-// -----------------------------------------------------------------------------------------------
-// Hypothesis 6: Behavior during the swap window (requests arriving mid-swap).
-// -----------------------------------------------------------------------------------------------
-
-// TestReanchorPoC_H6_AnchorHoldsUntilNewClientReady is the inversion of the original H6 finding.
-// The originally-observed broken behavior: reconcileCredentials flipped the rotator's anchor
-// synchronously and then built the new client asynchronously, opening a window where GetClient()
-// (== clients[AnchorKey()]) returned nil for the not-yet-registered new key. The fix builds the new
-// client synchronously and only commits the anchor once it reports Initialized, so no such nil window
-// exists: while the new client is still building, the anchor stays on the old key and GetClient() keeps
-// returning the old, still-serving client.
-func TestReanchorPoC_H6_AnchorHoldsUntilNewClientReady(t *testing.T) {
-	envConfig := st.EnvMain.Config
-
-	mockLog := ldlogtest.NewMockLog()
-	defer mockLog.DumpIfTestFailed(t)
-
-	clientCh := make(chan *testclient.FakeLDClient, 10)
-	inner := testclient.FakeLDClientFactoryWithChannel(true, clientCh)
-
-	// gate blocks construction of the NEW anchor's client so we can observe the swap window
-	// deterministically. entered signals that the synchronous build has reached the factory (and is now
-	// blocked on gate) — at which point the re-anchor is mid-flight but has not yet committed.
-	gate := make(chan struct{})
-	entered := make(chan struct{}, 1)
-	gatedFactory := func(sdkKey config.SDKKey, cfg ld.Config, timeout time.Duration) (sdks.LDClientContext, error) {
-		if sdkKey == reanchorTestKey2 {
-			entered <- struct{}{}
-			<-gate
-		}
-		return inner(sdkKey, cfg, timeout)
-	}
-
-	readyCh := make(chan EnvContext, 1)
-	env := makeBasicEnv(t, envConfig, gatedFactory, mockLog.Loggers, readyCh)
-	defer env.Close()
-
-	require.Equal(t, env, requireEnvReady(t, readyCh))
-	client1 := requireClientReady(t, clientCh)
-	require.Eventually(t, func() bool { return env.GetClient() == client1 }, time.Second, 10*time.Millisecond)
-
-	// The re-anchor is synchronous, so the reconcile blocks in the gated factory until we release it.
-	// Drive it on a background goroutine and build the set up front (keeping require off that goroutine).
-	start := time.Unix(1000, 0)
-	set, err := credential.NewAcceptedSetBuilder().
-		WithAnchor(credential.SDKKeyParams{Value: reanchorTestKey2}).
-		WithSDKKey(credential.SDKKeyParams{Value: envConfig.SDKKey, Expiry: util.PtrOrNil(start.Add(time.Hour))}).
-		Build()
-	require.NoError(t, err)
-	done := make(chan struct{})
-	go func() {
-		defer close(done)
-		env.(*envContextImpl).reconcileCredentials(set, start)
-	}()
-
-	// The build has reached the factory and is blocked: the re-anchor is mid-flight, pre-commit.
-	<-entered
-	envImpl := env.(*envContextImpl)
-	assert.Same(t, client1, env.GetClient(), "GetClient() keeps returning the old client during the build — never nil")
-	assert.Equal(t, envConfig.SDKKey, envImpl.keyRotator.AnchorKey(), "the anchor stays on the old key until the new client is ready")
-
-	// Release the gate; the new client initializes, the anchor commits, and GetClient() advances.
-	close(gate)
-	<-done
-	client2 := requireClientReady(t, clientCh)
-	require.Eventually(t, func() bool { return env.GetClient() == client2 }, time.Second, 10*time.Millisecond,
-		"GetClient() returns the new client once it is registered and the anchor commits")
-	assert.Equal(t, reanchorTestKey2, envImpl.keyRotator.AnchorKey())
-}
-
-// -----------------------------------------------------------------------------------------------
-// Hypothesis 7: Failure modes — new client init fails.
-// -----------------------------------------------------------------------------------------------
-
-// TestReanchorPoC_H7_FailedNewClientRollsBackToOldAnchor is the inversion of the original H7 finding.
-// The originally-observed broken behavior: a failed new-client init left the anchor already flipped to
-// a key with no client, so GetClient() returned nil and the environment was broken even though the old
-// client was still alive. The fix builds and validates the new client before committing: on init
-// failure it does NOT commit the anchor, surfaces the error, and leaves the old anchor authoritative
-// (its client keeps serving). This is the all-or-nothing atomicity requirement applied to re-anchor.
-func TestReanchorPoC_H7_FailedNewClientRollsBackToOldAnchor(t *testing.T) {
-	envConfig := st.EnvMain.Config
-	fakeErr := errors.New("new anchor client failed to initialize")
-
-	mockLog := ldlogtest.NewMockLog()
-	defer mockLog.DumpIfTestFailed(t)
-
-	clientCh := make(chan *testclient.FakeLDClient, 10)
-	inner := testclient.FakeLDClientFactoryWithChannel(true, clientCh)
-
-	// Succeed for the original anchor; fail for the new anchor.
-	failingFactory := func(sdkKey config.SDKKey, cfg ld.Config, timeout time.Duration) (sdks.LDClientContext, error) {
-		if sdkKey == reanchorTestKey2 {
-			return nil, fakeErr
-		}
-		return inner(sdkKey, cfg, timeout)
-	}
-
-	readyCh := make(chan EnvContext, 1)
-	env := makeBasicEnv(t, envConfig, failingFactory, mockLog.Loggers, readyCh)
-	defer env.Close()
-
-	require.Equal(t, env, requireEnvReady(t, readyCh))
-	client1 := requireClientReady(t, clientCh)
-	require.Eventually(t, func() bool { return env.GetClient() == client1 }, time.Second, 10*time.Millisecond)
-
-	// Re-anchor onto a key whose client init fails (old key kept valid for a grace hour). The re-anchor
-	// is synchronous, so the init failure and rollback are complete by the time reanchor returns.
-	start := time.Unix(1000, 0)
-	reanchor(t, env, reanchorTestKey2, envConfig.SDKKey, start)
-
-	// Post-fix: the re-anchor rolled back. The env stays healthy on the old anchor, so GetInitError
-	// stays nil — setting it would 401 a still-serving env at the request middleware. The failure
-	// surfaces via a structured Error log instead. The anchor pointer stayed on the old key, whose
-	// client is still alive and serving, so GetClient() never returns nil and no client is installed
-	// for the failed new anchor.
-	assert.NoError(t, env.GetInitError(), "a failed re-anchor must not mark the still-serving env as failed")
-	mockLog.AssertMessageMatch(t, true, ldlog.Error, "Re-anchor to SDK key .* failed")
-	assert.Same(t, client1, env.GetClient(), "GetClient() still returns the old anchor's client after rollback")
-	envImpl := env.(*envContextImpl)
-	assert.Equal(t, envConfig.SDKKey, envImpl.keyRotator.AnchorKey(), "the anchor pointer stays on the old key")
-	envImpl.mu.RLock()
-	_, oldStillPresent := envImpl.clients[envConfig.SDKKey]
-	_, newInstalled := envImpl.clients[reanchorTestKey2]
-	envImpl.mu.RUnlock()
-	assert.True(t, oldStillPresent, "the old anchor's client is preserved")
-	assert.False(t, newInstalled, "no client is installed for the failed new anchor")
-}
